+++ core/src/runtime/record.rs
@@ -147,7 +147,7 @@ pub struct ExecutionRecord {
     /// The public values.
     pub public_values: PublicValues<u32, u32>,
 
-    pub nonce_lookup: HashMap<usize, u32>,
+    pub nonce_lookup: HashMap<u128, u32>,
 }
 
 // Event lenses connect the record to the events relative to a particular chip
