+++ core/src/syscall/precompiles/sha256/extend/mod.rs
@@ -11,7 +11,7 @@ use crate::runtime::{MemoryReadRecord, MemoryWriteRecord};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ShaExtendEvent {
-    pub lookup_id: usize,
+    pub lookup_id: u128,
     pub shard: u32,
     pub channel: u32,
     pub clk: u32,
